Bitflag ints changed to uint32_t
Minor oversight that I never really considered - the program expects all
regular "int" variables to be 32 bit integers, since that what it is by
default on MY system. I should have been explicitly stating that the
"flags" field on most of my objects is an unsigned, 32 bit integer.
Fixed in all cases.

+++ include/solitaire.h
@@ -30,7 +30,7 @@ typedef struct {
     Deck *fromref; // A reference to where a move originates
     Deck *toref; // A reference to where the move is going
     char *msg; // String pointer for messages
-    uint8_t flags; // GameFlags defined in flags.h
+    uint32_t flags; // GameFlags defined in flags.h
     int score; // Current game score
 } Solitaire;
